+++ tests/test_recreation.py
@@ -14,12 +14,14 @@ import json
 import queue
 import multiprocessing
 
-import Pyro4
 import numpy
-import pandas
 from osgeo import gdal
 from osgeo import ogr
 from osgeo import osr
+import pandas
+import pygeoprocessing
+import Pyro4
+import shapely
 import taskgraph
 import warnings
 
@@ -459,7 +461,7 @@ class TestRecServer(unittest.TestCase):
             self.assertTrue(len(ws) == 0)
 
     @_timeout(30.0)
-    def test_regression_local_server(self):
+    def test_execute_local_server(self):
         """Recreation base regression test on sample data on local server.
 
         Executes Recreation model all the way through scenario prediction.
@@ -780,6 +782,62 @@ class RecreationRegressionTests(unittest.TestCase):
         # Assert that target file was written and it is an empty dictionary
         assert(len(predictor_results) == 0)
 
+    def test_overlapping_features_in_polygon_predictor(self):
+        """Recreation test overlapping predictor features not double-counted.
+
+        If a polygon predictor contains features that overlap, the overlapping
+        area should only be counted once when calculating `polygon_area_coverage`
+        or `polygon_percent_coverage`.
+        """
+        from natcap.invest.recreation import recmodel_client
+
+        response_vector_path = os.path.join(self.workspace_dir, 'aoi.geojson')
+        response_polygons_pickle_path = os.path.join(
+            self.workspace_dir, 'response.pickle')
+        predictor_vector_path = os.path.join(
+            self.workspace_dir, 'predictor.geojson')
+        predictor_target_path = os.path.join(
+            self.workspace_dir, 'predictor.json')
+
+        srs = osr.SpatialReference()
+        srs.ImportFromEPSG(32610)  # a UTM system
+
+        # A unit square
+        response_geom = shapely.geometry.Polygon(
+            ((0., 0.), (0., 1.), (1., 1.), (1., 0.), (0., 0.)))
+        pygeoprocessing.shapely_geometry_to_vector(
+            [response_geom],
+            response_vector_path,
+            srs.ExportToWkt(),
+            'GEOJSON')
+
+        # Two overlapping polygons, including a unit square
+        predictor_geom_list = [
+            shapely.geometry.Polygon(
+                ((0., 0.), (0., 1.), (1., 1.), (1., 0.), (0., 0.))),
+            shapely.geometry.Polygon(
+                ((0., 0.), (0., 0.5), (0.5, 0.5), (0.5, 0.), (0., 0.)))]
+        pygeoprocessing.shapely_geometry_to_vector(
+            predictor_geom_list,
+            predictor_vector_path,
+            srs.ExportToWkt(),
+            'GEOJSON')
+
+        recmodel_client._prepare_response_polygons_lookup(
+            response_vector_path, response_polygons_pickle_path)
+        recmodel_client._polygon_area(
+            'polygon_area_coverage',
+            response_polygons_pickle_path,
+            predictor_vector_path,
+            predictor_target_path)
+
+        with open(predictor_target_path, 'r') as file:
+            data = json.load(file)
+            print(data)
+        actual_value = list(data.values())[0]
+        expected_value = 1
+        self.assertEqual(actual_value, expected_value)
+
     def test_least_squares_regression(self):
         """Recreation regression test for the least-squares linear model."""
         from natcap.invest.recreation import recmodel_client
@@ -821,7 +879,7 @@ class RecreationRegressionTests(unittest.TestCase):
             numpy.testing.assert_allclose(results[key], expected_results[key])
 
     @unittest.skip("skipping to avoid remote server call (issue #3753)")
-    def test_base_regression(self):
+    def test_base_execute(self):
         """Recreation base regression test on fast sample data.
 
         Executes Recreation model with default data and default arguments.
@@ -851,7 +909,7 @@ class RecreationRegressionTests(unittest.TestCase):
             os.path.join(args['workspace_dir'], 'scenario_results.shp'),
             os.path.join(REGRESSION_DATA, 'scenario_results_40000.csv'))
 
-    def test_square_grid_regression(self):
+    def test_square_grid(self):
         """Recreation square grid regression test."""
         from natcap.invest.recreation import recmodel_client
 
@@ -868,7 +926,7 @@ class RecreationRegressionTests(unittest.TestCase):
         utils._assert_vectors_equal(
             out_grid_vector_path, expected_grid_vector_path)
 
-    def test_hex_grid_regression(self):
+    def test_hex_grid(self):
         """Recreation hex grid regression test."""
         from natcap.invest.recreation import recmodel_client
 
@@ -886,8 +944,8 @@ class RecreationRegressionTests(unittest.TestCase):
             out_grid_vector_path, expected_grid_vector_path)
 
     @unittest.skip("skipping to avoid remote server call (issue #3753)")
-    def test_no_grid_regression(self):
-        """Recreation base regression on ungridded AOI."""
+    def test_no_grid_execute(self):
+        """Recreation execute on ungridded AOI."""
         from natcap.invest.recreation import recmodel_client
 
         args = {